move test class to better location
